==> 1.1 Junção Bases de Dados.ipynb <==
# <font color='darkred'>Tese de Mestrado</font>
## Tratamento de Dados - Junção de Datasets 

# Importação de bibliotecas
import pandas as pd
import warnings
warnings.filterwarnings("ignore")

 ### Leitura de Ficheiros e Eliminação de Duplicados
 - Ficheiro de 2009

df09 = pd.read_excel('Trab_Emp_2009.xlsx')
df_09_s_dupli = df09.drop_duplicates()
df_09_s_dupli.head()

df09.shape

df_09_s_dupli.shape

- Ficheiro de 2010

df10 = pd.read_excel('Trab_Emp_2010.xlsx')
df_10_s_dupli = df10.drop_duplicates()
df_10_s_dupli.head()

df10.shape

df_10_s_dupli.shape

- Ficheiro de 2011

df11 = pd.read_excel('Trab_Emp_2011.xlsx')
df_11_s_dupli = df11.drop_duplicates()

df11.shape

df_11_s_dupli.shape

- Ficheiro de 2012

df12 = pd.read_excel('Trab_Emp_2012.xlsx')
df_12_s_dupli = df12.drop_duplicates()

df12.shape

df_12_s_dupli.shape

- Ficheiro de 2013

df13 = pd.read_excel('Trab_Emp_2013.xlsx')
df_13_s_dupli = df13.drop_duplicates()

df13.shape

df_13_s_dupli.shape

- Ficheiro de 2014

df14 = pd.read_excel('Trab_Emp_2014.xlsx')
df_14_s_dupli = df14.drop_duplicates()

df14.shape

df_14_s_dupli.shape

- Ficheiro de 2015

df15 = pd.read_excel('Trab_Emp_2015.xlsx')
df_15_s_dupli = df15.drop_duplicates()

df15.shape

df_15_s_dupli.shape

- Ficheiro de 2016

df16 = pd.read_excel('Trab_Emp_2016.xlsx')
df_16_s_dupli = df16.drop_duplicates()

df16.shape

df_16_s_dupli.shape

- Ficheiro de 2017

df17 = pd.read_excel('Trab_Emp_2017.xlsx')
df_17_s_dupli = df17.drop_duplicates()

df17.shape

df_17_s_dupli.shape

- Ficheiro de 2018

df18 = pd.read_excel('Trab_Emp_2018.xlsx')
df_18_s_dupli = df18.drop_duplicates()

df18.shape

df_18_s_dupli.shape

- Ficheiro de 2019

df19 = pd.read_excel('Trab_Emp_2019.xlsx')
df_19_s_dupli = df19.drop_duplicates()

df19.shape

df_19_s_dupli.shape

df_09 = df_09_s_dupli
df_10 = df_10_s_dupli
df_11 = df_11_s_dupli 
df_12 = df_12_s_dupli
df_13 = df_13_s_dupli
df_14 = df_14_s_dupli
df_15 = df_15_s_dupli
df_16 = df_16_s_dupli
df_17 = df_17_s_dupli
df_18 = df_18_s_dupli
df_19 = df_19_s_dupli

#### Análise das irregularidades entre ficheiros
- O ficheiro referente a **2009** contém **58 colunas** enquanto que os ficheiros de **2010 a 2019** contêm **67 colunas**;
- A maior divergência no nome das colunas ocorre entre 2009 e os restantes ficheiros;
- A partir de **2016** as colunas **caem2 e caem4_COD** (2010-2015) passam a chamar-se **CAE2 e CAE4_COD**;
- Os anos **2017 a 2019** passam a chamar-se **'ano_constituicao'**;
- No ano de **2019** as variáveis **'data_adm', 'data_prom', 'IRCT_Label'** estão **diferentes** dos restantes anos.

### Tratamento e Uniformização entre Ficheiros
- **Uniformização do nome** das colunas do ficheiro de **2009** face aos restantes ficheiros.

# Uniformização do nome das colunas em comparação com os dados a partir de 2010

df_09.columns = df_09.columns.str.replace('_09', '')
df_09.columns = df_09.columns.str.replace('09', '')
df_09.columns = df_09.columns.str.replace('n1emp', 'nut1_emp')
df_09.columns = df_09.columns.str.replace('n2emp', 'nut2_emp')
df_09.columns = df_09.columns.str.replace('vvend', 'vn')
df_09.columns = df_09.columns.str.replace('irc', 'irct')
df_09.columns = df_09.columns.str.replace('idade_TB_COD', 'idade_Cod')
df_09.columns = df_09.columns.str.replace('antg', 'antig')
df_09.columns = df_09.columns.str.replace('nqua1', 'nqual1')
df_09.columns = df_09.columns.str.replace('dtadm', 'dt_adm')
df_09.columns = df_09.columns.str.replace('dtulp', 'dt_prom')
df_09.columns = df_09.columns.str.replace('stpro', 'sitpro')
df_09.columns = df_09.columns.str.replace('ctcont', 'tipo_contr1')
df_09.columns = df_09.columns.str.replace('rprg', 'prest_reg')
df_09.columns = df_09.columns.str.replace('rpirg', 'prest_irreg')
df_09.columns = df_09.columns.str.replace('rganh', 'rganho')
df_09.columns = df_09.columns.str.replace('nhnor', 'hnormais')
df_09.columns = df_09.columns.str.replace('nhext', 'hextra')
df_09.columns = df_09.columns.str.replace('npnor', 'pnt')
df_09.columns = df_09.columns.str.replace('caem2', 'cae2')
df_09.columns = df_09.columns.str.replace('NPC_FIC', 'npc_fic')

print(df_09.columns)

df_09.shape

df_09.head()

- **Uniformização do nome** das colunas do ficheiro de **2010** face aos restantes ficheiros.

# Conerter todos os nomes das tabelas que estão em Caps Lock para minúsculas
def convert_to_lowercase(column_name):
    if isinstance(column_name, str) and column_name.isupper():
        return column_name.lower()
    else:
        return column_name

# Aplicar a função às colunas
df_10.columns = df_10.columns.map(convert_to_lowercase)
df_11.columns = df_11.columns.map(convert_to_lowercase)
df_12.columns = df_12.columns.map(convert_to_lowercase)
df_13.columns = df_13.columns.map(convert_to_lowercase)
df_14.columns = df_14.columns.map(convert_to_lowercase)
df_15.columns = df_15.columns.map(convert_to_lowercase)
df_16.columns = df_16.columns.map(convert_to_lowercase)
df_17.columns = df_17.columns.map(convert_to_lowercase)
df_18.columns = df_18.columns.map(convert_to_lowercase)
df_19.columns = df_19.columns.map(convert_to_lowercase)

df_10.columns = df_10.columns.str.replace('caem1', 'caem1l')

df_10.shape

print(df_10.columns)

df_17.columns = df_17.columns.str.replace('ano_constituicao', 'ancon')
df_18.columns = df_18.columns.str.replace('ano_constituicao', 'ancon')
df_19.columns = df_19.columns.str.replace('ano_constituicao', 'ancon')
df_19.columns = df_19.columns.str.replace('data_adm', 'dt_adm')
df_19.columns = df_19.columns.str.replace('data_prom', 'dt_prom')
df_19.columns = df_19.columns.str.replace('IRCT_Label', 'irct')

#### Tratamento da variável idade para aparecer o ano de nascimento

ano_referência = 2009 
df_09['ano_nascimento'] = ano_referência-df_09['idade_Cod']

df_09.head()

df_09.shape

### Filtragem ntrab pelo 0

dataframes = [df_09, df_10, df_11, df_12, df_13, df_14, df_15, df_16, df_17, df_18, df_19] 

def filter_ntrab(df):
    return df[df['ntrab'] != 0]

df_names = [f'df_{str(i).zfill(2)}' for i in range(9,20)]

for name in df_names:
    globals()[name] = filter_ntrab(globals()[name])

for i in df_names:
    print(f"{name} filtrado:")
    print(globals()[name])

df_09.shape

df_10.shape

df_11.shape

df_12.shape

df_13.shape

df_14.shape

df_15.shape

df_16.shape

df_17.shape

df_18.shape

df_19.shape

### Ntrab único em cada DataFrame

def concat_values(series):
    return '|'.join(map(str,series))

df_09_unique = df_09.groupby('ntrab').agg(concat_values).reset_index()
df_09_unique.shape

df_10_unique = df_10.groupby('ntrab').agg(concat_values).reset_index()
df_10_unique.shape

df_11_unique = df_11.groupby('ntrab').agg(concat_values).reset_index()
df_11_unique.shape

df_12_unique = df_12.groupby('ntrab').agg(concat_values).reset_index()
df_12_unique.shape

df_13_unique = df_13.groupby('ntrab').agg(concat_values).reset_index()
df_13_unique.shape

df_14_unique = df_14.groupby('ntrab').agg(concat_values).reset_index()
df_14_unique.shape

df_15_unique = df_15.groupby('ntrab').agg(concat_values).reset_index()
df_15_unique.shape

df_16_unique = df_16.groupby('ntrab').agg(concat_values).reset_index()
df_16_unique.shape

df_17_unique = df_17.groupby('ntrab').agg(concat_values).reset_index()
df_17_unique.shape

df_18_unique = df_18.groupby('ntrab').agg(concat_values).reset_index()
df_18_unique.shape

df_19_unique = df_19.groupby('ntrab').agg(concat_values).reset_index()
df_19_unique.shape

### Filtragem das Colunas de Análise 
As colunas de foco ao longo desta análise serão:
- **'tipo_contr'**, 'habil2', 'nuemp', 'cae2','tipo_contr1', 'rganho', 'hnormais', 'hextra', 'ntrab'

O ano de referência será 2009, por ser o primeiro ano de análise. Dado isto, serão consideradas mais algumas variáveis, tais como:
- 'sexo', 'ano_nascimento'

colunas_2009 = ['ntrab', 'sexo', 'ano_nascimento', 'habil1',  'tipo_contr1', 'nuemp', 'caem1l', 'prof_3d', 'rganho', 'hnormais', 'hextra']
df_09_final = df_09_unique[colunas_2009]
nomes_certos = {coluna: coluna if coluna == 'ntrab' else f"{coluna}_09" for coluna in df_09_final.columns}
df_09_final = df_09_final.rename(columns = nomes_certos)
df_09_final.head()

colunas_2010_2019 = ['ntrab', 'habil1', 'tipo_contr1', 'nuemp', 'caem1l', 'prof_3d', 'rganho', 'hnormais', 'hextra']
df_10_final = df_10_unique[colunas_2010_2019]
nomes_certos = {coluna: coluna if coluna == 'ntrab' else f"{coluna}_10" for coluna in df_10_final.columns}
df_10_final = df_10_final.rename(columns = nomes_certos)
df_10_final.head()

df_11_final = df_11_unique[colunas_2010_2019]
nomes_certos = {coluna: coluna if coluna == 'ntrab' else f"{coluna}_11" for coluna in df_11_final.columns}
df_11_final = df_11_final.rename(columns = nomes_certos)
df_11_final.head()

df_12_final = df_12_unique[colunas_2010_2019]
nomes_certos = {coluna: coluna if coluna == 'ntrab' else f"{coluna}_12" for coluna in df_12_final.columns}
df_12_final = df_12_final.rename(columns = nomes_certos)
df_12_final.head()

df_13_final = df_13_unique[colunas_2010_2019]
nomes_certos = {coluna: coluna if coluna == 'ntrab' else f"{coluna}_13" for coluna in df_13_final.columns}
df_13_final = df_13_final.rename(columns = nomes_certos)
df_13_final.head()

df_14_final = df_14_unique[colunas_2010_2019]
nomes_certos = {coluna: coluna if coluna == 'ntrab' else f"{coluna}_14" for coluna in df_14_final.columns}
df_14_final = df_14_final.rename(columns = nomes_certos)
df_14_final.head()

df_15_final = df_15_unique[colunas_2010_2019]
nomes_certos = {coluna: coluna if coluna == 'ntrab' else f"{coluna}_15" for coluna in df_15_final.columns}
df_15_final = df_15_final.rename(columns = nomes_certos)
df_15_final.head()

df_16_final = df_16_unique[colunas_2010_2019]
nomes_certos = {coluna: coluna if coluna == 'ntrab' else f"{coluna}_16" for coluna in df_16_final.columns}
df_16_final = df_16_final.rename(columns = nomes_certos)
df_16_final.head()

df_17_final = df_17_unique[colunas_2010_2019]
nomes_certos = {coluna: coluna if coluna == 'ntrab' else f"{coluna}_17" for coluna in df_17_final.columns}
df_17_final = df_17_final.rename(columns = nomes_certos)
df_17_final.head()

df_18_final = df_18_unique[colunas_2010_2019]
nomes_certos = {coluna: coluna if coluna == 'ntrab' else f"{coluna}_18" for coluna in df_18_final.columns}
df_18_final = df_18_final.rename(columns = nomes_certos)
df_18_final.head()

df_19_final = df_19_unique[colunas_2010_2019]
nomes_certos = {coluna: coluna if coluna == 'ntrab' else f"{coluna}_19" for coluna in df_19_final.columns}
df_19_final = df_19_final.rename(columns = nomes_certos)
df_19_final.head()

filtered_ntrab = df_09_final[df_09_final['nuemp_09'].str.count('\|')>1]['ntrab']
result_list_09 = filtered_ntrab.tolist()
len(result_list_09)

filtered_ntrab = df_10_final[df_10_final['nuemp_10'].str.count('\|')>1]['ntrab']
result_list_10 = filtered_ntrab.tolist()
len(result_list_10)

filtered_ntrab = df_11_final[df_11_final['nuemp_11'].str.count('\|')>1]['ntrab']
result_list_11 = filtered_ntrab.tolist()
len(result_list_11)

print(result_list_11)

filtered_ntrab = df_12_final[df_12_final['nuemp_12'].str.count('\|')>1]['ntrab']
result_list_12 = filtered_ntrab.tolist()
len(result_list_12)

filtered_ntrab = df_13_final[df_13_final['nuemp_13'].str.count('\|')>1]['ntrab']
result_list_13 = filtered_ntrab.tolist()
len(result_list_13)

filtered_ntrab = df_14_final[df_14_final['nuemp_14'].str.count('\|')>1]['ntrab']
result_list_14 = filtered_ntrab.tolist()
len(result_list_14)

filtered_ntrab = df_15_final[df_15_final['nuemp_15'].str.count('\|')>1]['ntrab']
result_list_15 = filtered_ntrab.tolist()
len(result_list_15)

filtered_ntrab = df_16_final[df_16_final['nuemp_16'].str.count('\|')>1]['ntrab']
result_list_16 = filtered_ntrab.tolist()
len(result_list_16)

filtered_ntrab = df_17_final[df_17_final['nuemp_17'].str.count('\|')>1]['ntrab']
result_list_17 = filtered_ntrab.tolist()
len(result_list_17)

filtered_ntrab = df_18_final[df_18_final['nuemp_18'].str.count('\|')>1]['ntrab']
result_list_18 = filtered_ntrab.tolist()
len(result_list_18)

filtered_ntrab = df_19_final[df_19_final['nuemp_19'].str.count('\|')>1]['ntrab']
result_list_19 = filtered_ntrab.tolist()
len(result_list_19)

### Exportação dos ficheiros excel finais

# Exportar o DataFrame para um arquivo CSV
df_09_final.to_excel('df_09.xlsx', index=False)
df_10_final.to_excel('df_10.xlsx', index=False)
df_11_final.to_excel('df_11.xlsx', index=False)
df_12_final.to_excel('df_12.xlsx', index=False)
df_13_final.to_excel('df_13.xlsx', index=False)
df_14_final.to_excel('df_14.xlsx', index=False)
df_15_final.to_excel('df_15.xlsx', index=False)
df_16_final.to_excel('df_16.xlsx', index=False)
df_17_final.to_excel('df_17.xlsx', index=False)
df_18_final.to_excel('df_18.xlsx', index=False)
df_19_final.to_excel('df_19.xlsx', index=False)

---
Beatriz Lapa - Tese de Mestrado
